+++ R/print_md_semfindr.R
@@ -17,8 +17,9 @@
 #' @param digits The number of digits after the decimal.
 #' Default is 3.
 #'
-#' @param first Numeric. If not `NULL`, the default, it print
+#' @param first Numeric. If not `NULL`, it prints
 #' only the first *k* cases, *k* equal to `first`.
+#' Default is 10.
 #'
 #' @param sort Logical. If `TRUE`, the default, the cases
 #' will be sorted based on Mahalanobis distance.
@@ -56,11 +57,11 @@
 #' @export
 
 print.md_semfindr <- function(x,
-                               digits = 3,
-                               first = NULL,
-                               sort = TRUE,
-                               decreasing = TRUE,
-                               ...) {
+                              digits = 3,
+                              first = 10,
+                              sort = TRUE,
+                              decreasing = TRUE,
+                              ...) {
     if (is.null(first)) {
         first <- nrow(x)
       }
@@ -88,7 +89,12 @@ print.md_semfindr <- function(x,
     cat("\nNote:\n")
 
     if (first != nrow(x)) {
-        cat("- Only the first", first, "cases are displayed.\n")
+        cat("- Only the first ",
+            first,
+            " case(s) is/are displayed.",
+            " Set ", sQuote("first"),
+            " to NULL to display all cases.",
+            "\n", sep = "")
       } else {
         cat("- All stored cases are displayed.\n")
       }